+++ src/append/file.rs
@@ -2,11 +2,11 @@
 //!
 //! Requires the `file_appender` feature.
 
-use derivative::Derivative;
 use log::Record;
 use parking_lot::Mutex;
 use std::{
     fs::{self, File, OpenOptions},
+    fmt::{Debug, Formatter},
     io::{self, BufWriter, Write},
     path::{Path, PathBuf},
 };
@@ -32,13 +32,20 @@ pub struct FileAppenderConfig {
 }
 
 /// An appender which logs to a file.
-#[derive(Derivative)]
-#[derivative(Debug)]
 pub struct FileAppender {
     path: PathBuf,
-    #[derivative(Debug = "ignore")]
     file: Mutex<SimpleWriter<BufWriter<File>>>,
     encoder: Box<dyn Encode>,
+}
+
+impl Debug for FileAppender {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct(stringify!(FileAppender))
+            .field("path", &self.path)
+            // ignore file
+            .field("encoder", &self.encoder)
+            .finish()
+    }
 }
 
 impl Append for FileAppender {